Split offline and AWS editions of the integration tests
